i18n: use full Thai month names everywhere — no abbreviations (#241)
Owner 2026-09-13: "ก.ย." → "กันยายน" across the whole system. Replaced
every abbreviated Thai month source with full names:

- Abbreviated month arrays → full names: doctor-fee rounds + notify,
  accounta (company / recurring / credit-cards), ascenda (daily /
  revenue), mounjaro (clinical / patient detail), financial-analysis.
- Intl formatters month:"short" → month:"long": doctor-fee overview,
  delivera COD, IR reports (list + detail), payroll PT breakdown.
- attendance shortMonth: full name + full year (was sliced 3 chars).
- credit-cards / ascenda-daily: full year alongside the full month.
- revshare TH_MONTHS_ABBR now aliases TH_MONTHS_FULL so any leftover
  reference still renders the full name.
- Employee-export placeholder text: "ณ ก.ย." → "ณ กันยายน".

Weekday abbreviations are left as-is (not requested). tsc/build/i18n/
risks green; df tests 112/112.

+++ src/app/admin/delivera/report/CodClient.tsx
@@ -9,7 +9,7 @@ type CodItem = { order_id: number; order_no: string; rider_name: string | null; 
 function fmtWhen(iso: string | null): string {
   if (!iso) return "";
   const d = new Date(iso.replace(" ", "T") + "Z");
-  return isNaN(d.getTime()) ? iso : d.toLocaleString("th-TH", { day: "numeric", month: "short", hour: "2-digit", minute: "2-digit", timeZone: "Asia/Bangkok" });
+  return isNaN(d.getTime()) ? iso : d.toLocaleString("th-TH", { day: "numeric", month: "long", hour: "2-digit", minute: "2-digit", timeZone: "Asia/Bangkok" });
 }
 
 export default function CodClient() {
