+++ src/app/admin/persona/doctor-fee/DoctorFeeClient.tsx
@@ -12,7 +12,7 @@ function pct(rate: number): string {
 }
 function fmtDay(date: string): string {
   const d = new Date(`${date}T00:00:00+07:00`);
-  return d.toLocaleDateString("th-TH", { weekday: "short", day: "numeric", month: "short", timeZone: "Asia/Bangkok" });
+  return d.toLocaleDateString("th-TH", { weekday: "short", day: "numeric", month: "long", timeZone: "Asia/Bangkok" });
 }
 function monthBounds(ym: string): { start: string; end: string } {
   const [y, m] = ym.split("-").map(Number);
